Move async_utils outside `gmeep`
