+++ gplugins/gmeep/write_sparameters_meep_mpi.py
@@ -198,7 +198,7 @@ def write_sparameters_meep_mpi(
     if live_output:
         import asyncio
 
-        from gplugins.gmeep.async_utils import execute_and_stream_output
+        from gplugins.async_utils import execute_and_stream_output
 
         asyncio.run(
             execute_and_stream_output(
